[hotfix][table-planner-blink] Move TableResultUtils into blink planner
